First round of fixes of several vulnarabilities
 - The vulnarabilities would enable unauthorized persons to modify data
   in your Xerte Toolkits installation

Many thanks to Bauke Gehem, System addin at Summa College, Eindhoven, The
Netherlands

+++ website_code/php/folders/delete_folder.php
@@ -30,6 +30,12 @@
 require_once('../../../config.php');
 include "../folder_library.php";
 
+if (!isset($_SESSION['toolkits_logon_username']))
+{
+    _debug("Session is invalid or expired");
+    die("Session is invalid or expired");
+}
+
 delete_folder($_POST['folder_id']);
 
 ?>
